refactor: Move utility functions to their own private module
