+++ eth_account/hdaccount/deterministic.py
@@ -37,41 +37,20 @@ Definitions
 #   the other functions (as Ethereum uses an Account-based system)
 # - Unlike other libraries, this library does not use Bitcoin key serialization, because it is
 #   not intended to be ultimately used for Bitcoin key derivations. This presents a simplified API.
-import hashlib
-import hmac
 from typing import (
     Tuple,
     Union,
 )
 
-from eth_keys import (
-    keys,
-)
 from eth_utils import (
     to_int,
 )
-from hexbytes import (
-    HexBytes,
+
+from ._utils import (
+    SECP256K1_N,
+    ec_point,
+    hmac_sha512,
 )
-
-SECP256K1_N = int("FFFFFFFF_FFFFFFFF_FFFFFFFF_FFFFFFFE_BAAEDCE6_AF48A03B_BFD25E8C_D0364141", 16)
-
-
-def hmac_sha512(chain_code: bytes, data: bytes) -> bytes:
-    """
-    As specified by RFC4231 - https://tools.ietf.org/html/rfc4231
-    """
-    return hmac.new(chain_code, data, hashlib.sha512).digest()
-
-
-def ec_point(pkey: bytes) -> bytes:
-    """
-    Compute `point(p)`, where `point` is ecdsa point multiplication
-
-    Note: Result is ecdsa public key serialized to compressed form
-    """
-    return keys.PrivateKey(HexBytes(pkey)).public_key.to_compressed_bytes()
-
 
 class SoftNode(int):
     """
